# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CredX API",
    "description": "AI-powered corporate credit underwriting: ingestion, research, fraud graph, explainable scoring, CAM, copilot.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/cases/{case_id}/notes": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "List Notes",
        "operationId": "list_notes_api_v1_cases__case_id__notes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NoteOut"
                  },
                  "title": "Response List Notes Api V1 Cases  Case Id  Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Create Note",
        "operationId": "create_note_api_v1_cases__case_id__notes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/preview-impact": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Preview Impact",
        "description": "Live 'what will this note do to the score' as the analyst types.",
        "operationId": "preview_impact_api_v1_notes_preview_impact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Preview Impact Api V1 Notes Preview Impact Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notes/{note_id}": {
      "patch": {
        "tags": [
          "workflow"
        ],
        "summary": "Update Note",
        "operationId": "update_note_api_v1_notes__note_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workflow"
        ],
        "summary": "Delete Note",
        "operationId": "delete_note_api_v1_notes__note_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/overrides": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "List Overrides",
        "operationId": "list_overrides_api_v1_cases__case_id__overrides_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OverrideOut"
                  },
                  "title": "Response List Overrides Api V1 Cases  Case Id  Overrides Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Request Override",
        "operationId": "request_override_api_v1_cases__case_id__overrides_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/overrides/{override_id}/review": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Review Override",
        "operationId": "review_override_api_v1_overrides__override_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "override_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Override Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/escalations": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "List Escalations",
        "operationId": "list_escalations_api_v1_cases__case_id__escalations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EscalationOut"
                  },
                  "title": "Response List Escalations Api V1 Cases  Case Id  Escalations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Escalate",
        "operationId": "escalate_api_v1_cases__case_id__escalations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalationIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalations/{escalation_id}/resolve": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Resolve",
        "operationId": "resolve_api_v1_escalations__escalation_id__resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "escalation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Escalation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EscalationIn": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Reason"
          },
          "to_role": {
            "type": "string",
            "enum": [
              "credit_manager",
              "admin"
            ],
            "title": "To Role",
            "default": "credit_manager"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "EscalationIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NoteIn": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 3,
            "title": "Body"
          },
          "category": {
            "type": "string",
            "enum": [
              "site_visit",
              "management",
              "operations",
              "financial",
              "collateral",
              "compliance",
              "general"
            ],
            "title": "Category",
            "default": "general"
          },
          "kind": {
            "type": "string",
            "enum": [
              "note",
              "comment"
            ],
            "title": "Kind",
            "default": "note"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "impact_points": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 40.0,
                "minimum": -40.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Points"
          },
          "impact_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Rationale"
          },
          "five_c": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "character",
                  "capacity",
                  "capital",
                  "collateral",
                  "conditions"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Five C"
          },
          "include_in_cam": {
            "type": "boolean",
            "title": "Include In Cam",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "NoteIn"
      },
      "NoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "impact_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Points"
          },
          "inferred_impact": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inferred Impact"
          },
          "impact_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Rationale"
          },
          "five_c": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Five C"
          },
          "include_in_cam": {
            "type": "boolean",
            "title": "Include In Cam"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "preview_impact": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Impact"
          },
          "preview_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Rationale"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "parent_id",
          "kind",
          "category",
          "body",
          "impact_points",
          "inferred_impact",
          "impact_rationale",
          "five_c",
          "include_in_cam",
          "pinned",
          "author",
          "created_at"
        ],
        "title": "NoteOut"
      },
      "NoteUpdate": {
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "impact_points": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 40.0,
                "minimum": -40.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Points"
          },
          "clear_impact": {
            "type": "boolean",
            "title": "Clear Impact",
            "default": false
          },
          "include_in_cam": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include In Cam"
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned"
          }
        },
        "type": "object",
        "title": "NoteUpdate"
      },
      "OverrideIn": {
        "properties": {
          "field": {
            "type": "string",
            "enum": [
              "decision",
              "recommended_amount",
              "suggested_rate",
              "credit_score"
            ],
            "title": "Field"
          },
          "new_value": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "New Value"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "field",
          "new_value",
          "reason"
        ],
        "title": "OverrideIn"
      },
      "OverrideOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "field": {
            "type": "string",
            "title": "Field"
          },
          "original_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Value"
          },
          "new_value": {
            "type": "string",
            "title": "New Value"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "requested_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewed_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "field",
          "original_value",
          "new_value",
          "reason",
          "status",
          "requested_by",
          "reviewed_by",
          "reviewed_at",
          "review_comment",
          "created_at"
        ],
        "title": "OverrideOut"
      },
...
      "ReviewIn": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "ReviewIn"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}